Rename trone source files to TeraRangerI2C
